src/requirements.txt
ezdxf>=1.1.0
pydantic>=2.5.0
pyyaml>=6.0.1
tqdm>=4.66.0
